fix: SRP에 따른 클래스 분리 (Calculator, ExpressionStack)
